modified attack methods to include attack type code - changed variables and return types to double - added 0.1 for strong attack, 0.2 for weak. - modified receiveAttack() to cast attackDamage as int to ignore decimal.

+++ src/com/ironhack/RPG/Model/Attacker.java
@@ -2,8 +2,8 @@ package com.ironhack.RPG.Model;
 
 public interface Attacker {
 
-    int attack();
+    double attack();
 
-    void receiveAttack(int attackDamage);
+    void receiveAttack(double attackDamage);
 
 }
